Factor out Array -> R vector code in separate file
